Move adjoint test setup to separate file
