+++ test_adjoint/test_utils.py
@@ -3,6 +3,7 @@ from goalie_adjoint import *
 from goalie.adjoint import annotate_qoi
 import numpy as np
 import unittest
+from setup_adjoint_tests import *
 
 
 class TestAdjointUtils(unittest.TestCase):
